fix: rename class
